+++ src/mnemonic_im.hpp
@@ -1,0 +1,18 @@
+#pragma once
+#include "common.h"
+#include "mnemonic.h"
+
+void mnemonic_IM(LineData* line)
+{
+    if (mnemonic_format_check(line, 2, TokenType::Numeric)) {
+        line->machine.push_back(0xED);
+        switch (atoi(line->token[1].second.c_str())) {
+            case 0: line->machine.push_back(0x46); break;
+            case 1: line->machine.push_back(0x56); break;
+            case 2: line->machine.push_back(0x5E); break;
+            default:
+                line->error = true;
+                line->errmsg = "Unsupported interrupt mode: " + line->token[1].second;
+        }
+    }
+}
